HMS-5948: refactor external repos command
Introduces cli library for external-repos commands, and moves the
command implementations to a separate package.

Consolidates the cleanup jobs into the same cron. Pulp orphan cleanup is
left on a separate weekly cadence, so pulp workers are not overwhelmed.
