+++ .github/workflows/playwright-actions.yml
@@ -248,7 +248,7 @@ jobs:
         run: OPTIONS_REPOSITORY_IMPORT_FILTER=small make repos-import
 
       - name: Run snapshot for redhat repo
-        run: go run cmd/external-repos/main.go snapshot https://cdn.redhat.com/content/dist/rhel9/9/aarch64/codeready-builder/os/
+        run: go run cmd/external-repos/main.go snapshot --url https://cdn.redhat.com/content/dist/rhel9/9/aarch64/codeready-builder/os/
 
       - name: Run API Playwright tests
         working-directory: _playwright-tests
